refactor: Search 와 Video 통합
